% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateFormula.R
\name{$.MultivariateFormula}
\alias{$.MultivariateFormula}
\title{$ operator for multivariate formula}
\usage{
\method{$}{MultivariateFormula}(f, a)
}
\arguments{
\item{f}{formula}

\item{a}{attribute}
}
\description{
S3 helper function to retrieve attributes as if it was named values
}
\keyword{internal}
