% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critConvergence.R
\name{critConvergence}
\alias{critConvergence}
\title{Auxiliary function for controlling SCGLR fitting}
\usage{
critConvergence(tol = 1e-06, maxit = 50)
}
\arguments{
\item{tol}{positive convergence threshold.}

\item{maxit}{integer, maximum number of iterations.}
}
\value{
a list containing elements named as the arguments.
}
\description{
Auxiliary function for scglr fitting used to
construct a convergence control argument.
}
