% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactorSCGLR.R
\name{FactorSCGLR}
\alias{FactorSCGLR}
\title{Function that fits the scglr model with factors}
\usage{
FactorSCGLR(
  formula,
  data,
  family,
  H = c(1, 1),
  J = 1,
  size = NULL,
  offset = NULL,
  crit = list(),
  method = methodSR()
)
}
\arguments{
\item{formula}{an object of class \code{MultivariateFormula} (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{a data frame to be modeled.}

\item{family}{a vector of character of the same length as the number of dependent variables:
"bernoulli", "binomial", "poisson" or "gaussian" is allowed.}

\item{H}{a vector of integer representing the number of components per theme.}

\item{J}{the number of factors}

\item{size}{describes the number of trials for the binomial dependent variables.}

\item{offset}{used for the poisson dependent variables.}

\item{crit}{a list of two elements : maxit and tol, describing respectively the maximum number of iterations and
the tolerance convergence criterion for the Expectation Maximization algorithm. Default is set to 50 and 10e-6 respectively.}

\item{method}{structural relevance criterion.}
}
\value{
\item{U}{the set of loading vectors}

\item{comp}{the set of components}

\item{eta}{the fitted linear predictor}

\item{coef}{contains regression parameters: intercept, gamma and delta}

\item{B}{the factors loading}

\item{G}{the set of factors}

\item{sigma2}{the residual variances for gaussian responses}

\item{offset}{the offset used for the poisson dependent variables}

\item{Z}{the set of working variables}

\item{Y}{the response matrix}

\item{Theme}{the list of standardized explanatory themes}

\item{A}{the supplementary explanatory variables}

\item{W}{the matrix of GLM weights}

\item{family}{the distributions}
}
\description{
Function that fits the scglr model with factors
}
\examples{
# load sample data
data <- genus
# get variable names from dataset
n <- names(data)
ny <- n[grep("^gen",n)]
nx1 <- n[grep("^evi",n)]
nx2 <- n[grep("^pluvio",n)]
na <- c("geology", "altitude", "forest", "lon", "lat")
# build multivariate formula
form <- multivariateFormula(Y = ny, X = list(nx1, nx2), A = na)
# define family
fam <- rep("poisson", length(ny))
# run function
H <- c(2,2)
J <- 2
met <- methodSR(l=4, s=0.5)
res <- FactorSCGLR(formula=form, data=data, H=H, J=J,
                   family=fam, method=met, offset = data$surface)

}
