% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InformationCriterion.R
\name{InformationCriterion}
\alias{InformationCriterion}
\title{Information criterion for scglr with factors}
\usage{
InformationCriterion(x)
}
\arguments{
\item{x}{an object from FactorsSCGLR}
}
\value{
\item{aic}{akaike information criterion}

\item{bic}{bayesian information criterion}
}
\description{
Information criterion for scglr with factors
}
\examples{
# load sample data
data <- genus
data <- as.data.frame(apply(data, 2, as.numeric ))
# get variable names from dataset
n <- names(data)
ny <- n[grep("^gen",n)]
nx1 <- n[grep("^evi",n)]
nx2 <- n[grep("^pluvio",n)]
na <- c("geology", "altitude", "forest", "lon", "lat")
# build multivariate formula
form <- multivariateFormula(Y = ny, X = list(nx1, nx2), A = na)
# define family
fam <- rep("poisson", length(ny))
# run function
H <- c(2,2)
J <- 2
met <- methodSR(l=4, s=0.5)
res <- FactorsSCGLR(formula=form, data=data, H=H, J=J,
                   family=fam, method=met, offset = data$surface)

# compute the information criterion
IC <- InformationCriterion(res)
IC$aic
IC$bic
}
