% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methodSR}
\alias{methodSR}
\title{Regularization criterion types}
\usage{
methodSR(
  phi = "vpi",
  l = 1,
  s = 1/2,
  maxiter = 1000,
  epsilon = 1e-06,
  bailout = 10
)
}
\arguments{
\item{phi}{character string describing structural relevance used in the regularization process.
Allowed values are "vpi" for Variable Powered Inertia and "cv" for Component Variance. Default to "vpi".}

\item{l}{is an integer argument (>=1) tuning the importance of variable bundle locality.}

\item{s}{is a numeric argument (between 0 and 1) tuning the strength of structural relevance with respect to goodness of fit.}

\item{maxiter}{integer for maximum number of iterations of SR function}

\item{epsilon}{positive convergence threshold}

\item{bailout}{integer argument}
}
\description{
Regularization criterion types
}
