% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateFormula.R
\name{print.MultivariateFormula}
\alias{print.MultivariateFormula}
\title{print a multivariate formula}
\usage{
\method{print}{MultivariateFormula}(x, ...)
}
\arguments{
\item{x}{a formula}

\item{...}{unused}
}
\description{
S3 helper function to print a multivariate formula
NB use $ semantic to retrieve metadata
}
\keyword{internal}
